requirements.txt
psycopg2-binary==2.9.9
kafka-python-ng==2.2.2